+++ js/react-shadcn/src/index.ts
@@ -1,5 +1,3 @@
-export * from './contract';
-export { createHttpClient, type HttpClientOptions } from './client';
-export { compileField, compileForm, serverOnlyFields } from './rules';
+export * from '@spawnflow/core';
 export { SpawnForm, type SpawnFormProps } from './SpawnForm';
 export { defaultWidgets, type WidgetComponent, type WidgetProps } from './widgets';